+++ client/src/routeTree.gen.ts
@@ -18,7 +18,6 @@ import { Route as F125RouteImport } from './routes/f125'
 import { Route as Fm23RouteImport } from './routes/fm23'
 import { Route as IracingRouteImport } from './routes/iracing'
 import { Route as GameLiveRouteImport } from './routes/$game/live'
-import { Route as GameidAnalyseRouteImport } from './routes/$gameid/analyse'
 import { Route as GameidCarsRouteImport } from './routes/$gameid/cars'
 import { Route as GameidChatsRouteImport } from './routes/$gameid/chats'
 import { Route as GameidCompareRouteImport } from './routes/$gameid/compare'
@@ -44,6 +43,8 @@ import { Route as PortableCombo1RouteImport } from './routes/portable.combo-1'
 import { Route as PortableCombo2RouteImport } from './routes/portable.combo-2'
 import { Route as GameidExperimentsIndexRouteImport } from './routes/$gameid/experiments.index'
 import { Route as GameidExperimentsExperimentIdRouteImport } from './routes/$gameid/experiments.$experimentId'
+import { Route as GameidSessionsIndexRouteImport } from './routes/$gameid/sessions.index'
+import { Route as GameidSessionsAnalyseRouteImport } from './routes/$gameid/sessions.analyse'
 import { Route as GameidTracksIndexRouteImport } from './routes/$gameid/tracks.index'
 import { Route as AcEvoSetupsIndexRouteImport } from './routes/ac-evo/setups/index'
 import { Route as AcEvoSetupsImportRouteImport } from './routes/ac-evo/setups/import'
@@ -112,11 +113,6 @@ const GameLiveRoute = GameLiveRouteImport.update({
   id: '/$game/live',
   path: '/$game/live',
   getParentRoute: () => rootRouteImport,
-} as any)
-const GameidAnalyseRoute = GameidAnalyseRouteImport.update({
-  id: '/analyse',
-  path: '/analyse',
-  getParentRoute: () => GameidRoute,
 } as any)
 const GameidCarsRoute = GameidCarsRouteImport.update({
   id: '/cars',
@@ -244,6 +240,16 @@ const GameidExperimentsExperimentIdRoute =
     path: '/$experimentId',
     getParentRoute: () => GameidExperimentsRoute,
   } as any)
+const GameidSessionsIndexRoute = GameidSessionsIndexRouteImport.update({
+  id: '/',
+  path: '/',
+  getParentRoute: () => GameidSessionsRoute,
+} as any)
+const GameidSessionsAnalyseRoute = GameidSessionsAnalyseRouteImport.update({
+  id: '/analyse',
+  path: '/analyse',
+  getParentRoute: () => GameidSessionsRoute,
+} as any)
 const GameidTracksIndexRoute = GameidTracksIndexRouteImport.update({
   id: '/',
   path: '/',
@@ -373,14 +379,13 @@ export interface FileRoutesByFullPath {
   '/fm23': typeof Fm23RouteWithChildren
   '/iracing': typeof IracingRouteWithChildren
   '/$game/live': typeof GameLiveRoute
-  '/$gameid/analyse': typeof GameidAnalyseRoute
   '/$gameid/cars': typeof GameidCarsRoute
   '/$gameid/chats': typeof GameidChatsRoute
   '/$gameid/compare': typeof GameidCompareRoute
   '/$gameid/driver': typeof GameidDriverRoute
   '/$gameid/experiments': typeof GameidExperimentsRouteWithChildren
   '/$gameid/raw': typeof GameidRawRoute
-  '/$gameid/sessions': typeof GameidSessionsRoute
+  '/$gameid/sessions': typeof GameidSessionsRouteWithChildren
   '/$gameid/tracks': typeof GameidTracksRouteWithChildren
   '/ac-evo/setups': typeof AcEvoSetupsRouteWithChildren
   '/acc/setups': typeof AccSetupsRouteWithChildren
@@ -398,6 +403,7 @@ export interface FileRoutesByFullPath {
   '/iracing/': typeof IracingIndexRoute
   '/portable/': typeof PortableIndexRoute
   '/$gameid/experiments/$experimentId': typeof GameidExperimentsExperimentIdRoute
+  '/$gameid/sessions/analyse': typeof GameidSessionsAnalyseRoute
   '/ac-evo/setups/import': typeof AcEvoSetupsImportRoute
   '/ac-evo/setups/new': typeof AcEvoSetupsNewRoute
   '/acc/setups/import': typeof AccSetupsImportRoute
@@ -410,6 +416,7 @@ export interface FileRoutesByFullPath {
   '/iracing/live/driver': typeof IracingLiveDriverRoute
   '/iracing/live/pit': typeof IracingLivePitRoute
   '/$gameid/experiments/': typeof GameidExperimentsIndexRoute
+  '/$gameid/sessions/': typeof GameidSessionsIndexRoute
   '/$gameid/tracks/': typeof GameidTracksIndexRoute
   '/ac-evo/setups/': typeof AcEvoSetupsIndexRoute
   '/acc/setups/': typeof AccSetupsIndexRoute
@@ -428,13 +435,11 @@ export interface FileRoutesByTo {
   '/$gameid': typeof GameidRouteWithChildren
   '/dev': typeof DevRoute
   '/$game/live': typeof GameLiveRoute
-  '/$gameid/analyse': typeof GameidAnalyseRoute
   '/$gameid/cars': typeof GameidCarsRoute
   '/$gameid/chats': typeof GameidChatsRoute
   '/$gameid/compare': typeof GameidCompareRoute
   '/$gameid/driver': typeof GameidDriverRoute
   '/$gameid/raw': typeof GameidRawRoute
-  '/$gameid/sessions': typeof GameidSessionsRoute
   '/fm23/live': typeof Fm23LiveRouteWithChildren
   '/iracing/live': typeof IracingLiveRouteWithChildren
   '/portable/combo-1': typeof PortableCombo1Route
@@ -446,6 +451,7 @@ export interface FileRoutesByTo {
   '/iracing': typeof IracingIndexRoute
   '/portable': typeof PortableIndexRoute
   '/$gameid/experiments/$experimentId': typeof GameidExperimentsExperimentIdRoute
+  '/$gameid/sessions/analyse': typeof GameidSessionsAnalyseRoute
   '/ac-evo/setups/import': typeof AcEvoSetupsImportRoute
   '/ac-evo/setups/new': typeof AcEvoSetupsNewRoute
   '/acc/setups/import': typeof AccSetupsImportRoute
@@ -458,6 +464,7 @@ export interface FileRoutesByTo {
   '/iracing/live/driver': typeof IracingLiveDriverRoute
   '/iracing/live/pit': typeof IracingLivePitRoute
   '/$gameid/experiments': typeof GameidExperimentsIndexRoute
+  '/$gameid/sessions': typeof GameidSessionsIndexRoute
   '/$gameid/tracks': typeof GameidTracksIndexRoute
   '/ac-evo/setups': typeof AcEvoSetupsIndexRoute
   '/acc/setups': typeof AccSetupsIndexRoute
@@ -482,14 +489,13 @@ export interface FileRoutesById {
   '/fm23': typeof Fm23RouteWithChildren
   '/iracing': typeof IracingRouteWithChildren
   '/$game/live': typeof GameLiveRoute
-  '/$gameid/analyse': typeof GameidAnalyseRoute
   '/$gameid/cars': typeof GameidCarsRoute
   '/$gameid/chats': typeof GameidChatsRoute
   '/$gameid/compare': typeof GameidCompareRoute
   '/$gameid/driver': typeof GameidDriverRoute
   '/$gameid/experiments': typeof GameidExperimentsRouteWithChildren
   '/$gameid/raw': typeof GameidRawRoute
-  '/$gameid/sessions': typeof GameidSessionsRoute
+  '/$gameid/sessions': typeof GameidSessionsRouteWithChildren
   '/$gameid/tracks': typeof GameidTracksRouteWithChildren
   '/ac-evo/setups': typeof AcEvoSetupsRouteWithChildren
   '/acc/setups': typeof AccSetupsRouteWithChildren
@@ -507,6 +513,7 @@ export interface FileRoutesById {
   '/iracing/': typeof IracingIndexRoute
   '/portable/': typeof PortableIndexRoute
   '/$gameid/experiments/$experimentId': typeof GameidExperimentsExperimentIdRoute
+  '/$gameid/sessions/analyse': typeof GameidSessionsAnalyseRoute
   '/ac-evo/setups/import': typeof AcEvoSetupsImportRoute
   '/ac-evo/setups/new': typeof AcEvoSetupsNewRoute
   '/acc/setups/import': typeof AccSetupsImportRoute
@@ -519,6 +526,7 @@ export interface FileRoutesById {
   '/iracing/live/driver': typeof IracingLiveDriverRoute
   '/iracing/live/pit': typeof IracingLivePitRoute
   '/$gameid/experiments/': typeof GameidExperimentsIndexRoute
+  '/$gameid/sessions/': typeof GameidSessionsIndexRoute
   '/$gameid/tracks/': typeof GameidTracksIndexRoute
   '/ac-evo/setups/': typeof AcEvoSetupsIndexRoute
   '/acc/setups/': typeof AccSetupsIndexRoute
@@ -544,7 +552,6 @@ export interface FileRouteTypes {
     | '/fm23'
     | '/iracing'
     | '/$game/live'
-    | '/$gameid/analyse'
     | '/$gameid/cars'
     | '/$gameid/chats'
     | '/$gameid/compare'
@@ -569,6 +576,7 @@ export interface FileRouteTypes {
     | '/iracing/'
     | '/portable/'
     | '/$gameid/experiments/$experimentId'
+    | '/$gameid/sessions/analyse'
     | '/ac-evo/setups/import'
     | '/ac-evo/setups/new'
     | '/acc/setups/import'
@@ -581,6 +589,7 @@ export interface FileRouteTypes {
     | '/iracing/live/driver'
     | '/iracing/live/pit'
     | '/$gameid/experiments/'
+    | '/$gameid/sessions/'
     | '/$gameid/tracks/'
     | '/ac-evo/setups/'
     | '/acc/setups/'
@@ -599,13 +608,11 @@ export interface FileRouteTypes {
     | '/$gameid'
     | '/dev'
     | '/$game/live'
-    | '/$gameid/analyse'
     | '/$gameid/cars'
     | '/$gameid/chats'
     | '/$gameid/compare'
     | '/$gameid/driver'
     | '/$gameid/raw'
-    | '/$gameid/sessions'
     | '/fm23/live'
     | '/iracing/live'
     | '/portable/combo-1'
@@ -617,6 +624,7 @@ export interface FileRouteTypes {
     | '/iracing'
     | '/portable'
     | '/$gameid/experiments/$experimentId'
+    | '/$gameid/sessions/analyse'
     | '/ac-evo/setups/import'
     | '/ac-evo/setups/new'
     | '/acc/setups/import'
@@ -629,6 +637,7 @@ export interface FileRouteTypes {
     | '/iracing/live/driver'
     | '/iracing/live/pit'
     | '/$gameid/experiments'
+    | '/$gameid/sessions'
     | '/$gameid/tracks'
     | '/ac-evo/setups'
     | '/acc/setups'
@@ -652,7 +661,6 @@ export interface FileRouteTypes {
     | '/fm23'
     | '/iracing'
     | '/$game/live'
-    | '/$gameid/analyse'
     | '/$gameid/cars'
     | '/$gameid/chats'
     | '/$gameid/compare'
@@ -677,6 +685,7 @@ export interface FileRouteTypes {
     | '/iracing/'
     | '/portable/'
     | '/$gameid/experiments/$experimentId'
+    | '/$gameid/sessions/analyse'
     | '/ac-evo/setups/import'
     | '/ac-evo/setups/new'
     | '/acc/setups/import'
@@ -689,6 +698,7 @@ export interface FileRouteTypes {
     | '/iracing/live/driver'
     | '/iracing/live/pit'
     | '/$gameid/experiments/'
+    | '/$gameid/sessions/'
     | '/$gameid/tracks/'
     | '/ac-evo/setups/'
     | '/acc/setups/'
@@ -782,13 +792,6 @@ declare module '@tanstack/react-router' {
       fullPath: '/$game/live'
       preLoaderRoute: typeof GameLiveRouteImport
       parentRoute: typeof rootRouteImport
-    }
-    '/$gameid/analyse': {
-      id: '/$gameid/analyse'
-      path: '/analyse'
-      fullPath: '/$gameid/analyse'
-      preLoaderRoute: typeof GameidAnalyseRouteImport
-      parentRoute: typeof GameidRoute
     }
     '/$gameid/cars': {
       id: '/$gameid/cars'
@@ -964,6 +967,20 @@ declare module '@tanstack/react-router' {
       fullPath: '/$gameid/experiments/$experimentId'
       preLoaderRoute: typeof GameidExperimentsExperimentIdRouteImport
       parentRoute: typeof GameidExperimentsRoute
+    }
+    '/$gameid/sessions/': {
+      id: '/$gameid/sessions/'
+      path: '/'
+      fullPath: '/$gameid/sessions/'
+      preLoaderRoute: typeof GameidSessionsIndexRouteImport
+      parentRoute: typeof GameidSessionsRoute
+    }
+    '/$gameid/sessions/analyse': {
+      id: '/$gameid/sessions/analyse'
+      path: '/analyse'
+      fullPath: '/$gameid/sessions/analyse'
+      preLoaderRoute: typeof GameidSessionsAnalyseRouteImport
+      parentRoute: typeof GameidSessionsRoute
     }
     '/$gameid/tracks/': {
       id: '/$gameid/tracks/'
@@ -1145,6 +1162,20 @@ const GameidExperimentsRouteChildren: GameidExperimentsRouteChildren = {
 const GameidExperimentsRouteWithChildren =
   GameidExperimentsRoute._addFileChildren(GameidExperimentsRouteChildren)
 
+interface GameidSessionsRouteChildren {
+  GameidSessionsAnalyseRoute: typeof GameidSessionsAnalyseRoute
+  GameidSessionsIndexRoute: typeof GameidSessionsIndexRoute
+}
+
+const GameidSessionsRouteChildren: GameidSessionsRouteChildren = {
+  GameidSessionsAnalyseRoute: GameidSessionsAnalyseRoute,
+  GameidSessionsIndexRoute: GameidSessionsIndexRoute,
+}
+
+const GameidSessionsRouteWithChildren = GameidSessionsRoute._addFileChildren(
+  GameidSessionsRouteChildren,
+)
+
 interface GameidTracksRouteChildren {
   GameidTracksIndexRoute: typeof GameidTracksIndexRoute
   GameidTracksTrackOrdinalTabRoute: typeof GameidTracksTrackOrdinalTabRoute
@@ -1162,26 +1193,24 @@ const GameidTracksRouteWithChildren = GameidTracksRoute._addFileChildren(
 )
 
 interface GameidRouteChildren {
-  GameidAnalyseRoute: typeof GameidAnalyseRoute
   GameidCarsRoute: typeof GameidCarsRoute
   GameidChatsRoute: typeof GameidChatsRoute
   GameidCompareRoute: typeof GameidCompareRoute
   GameidDriverRoute: typeof GameidDriverRoute
   GameidExperimentsRoute: typeof GameidExperimentsRouteWithChildren
   GameidRawRoute: typeof GameidRawRoute
-  GameidSessionsRoute: typeof GameidSessionsRoute
+  GameidSessionsRoute: typeof GameidSessionsRouteWithChildren
   GameidTracksRoute: typeof GameidTracksRouteWithChildren
 }
 
 const GameidRouteChildren: GameidRouteChildren = {
-  GameidAnalyseRoute: GameidAnalyseRoute,
   GameidCarsRoute: GameidCarsRoute,
   GameidChatsRoute: GameidChatsRoute,
   GameidCompareRoute: GameidCompareRoute,
   GameidDriverRoute: GameidDriverRoute,
   GameidExperimentsRoute: GameidExperimentsRouteWithChildren,
   GameidRawRoute: GameidRawRoute,
-  GameidSessionsRoute: GameidSessionsRoute,
+  GameidSessionsRoute: GameidSessionsRouteWithChildren,
   GameidTracksRoute: GameidTracksRouteWithChildren,
 }